refactor: use Printe
