--- internal/cmd/extensions/printer.go
@@ -1,37 +1,0 @@
-// Copyright Mia srl
-// SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package extensions
-
-import (
-	"strings"
-
-	"github.com/mia-platform/miactl/internal/resources/extensibility"
-)
-
-func printExtensionsList(extensions []*extensibility.Extension) string {
-	strBuilder := &strings.Builder{}
-	table := NewTable(strBuilder)
-	table.Header("ID", "Name", "Description")
-	for _, extension := range extensions {
-		table.Row(
-			extension.ExtensionID,
-			extension.Name,
-			extension.Description,
-		)
-	}
-	table.Print()
-	return strBuilder.String()
-}
